linux: make it easlier to compare dts and build mainline

+++ 04_build_linux.sh
@@ -7,20 +7,35 @@ CC=${CC:-"${wdir}/riscv-toolchain/bin/riscv64-linux-"}
 cd ./linux/
 
 if [ ! -f ./.patched ] ; then
-	patch -p1 < ../patches/linux/0001-Add-BeagleV-Fire-device-tree.patch
-	patch -p1 < ../patches/linux/0001-PCIe-Change-controller-and-bridge-base-address.patch
-	patch -p1 < ../patches/linux/0001-GPIO-Add-Microchip-CoreGPIO-driver.patch
-	patch -p1 < ../patches/linux/0001-ADC-Add-Microchip-MCP356X-driver.patch
-	patch -p1 < ../patches/linux/0001-Microchip-QSPI-Add-regular-transfers.patch
-	patch -p1 < ../patches/linux/0001-BeagleV-Fire-Add-printk-to-IM219-driver-for-board-te.patch
-	patch -p1 < ../patches/linux/0001-MMC-SPI-Hack-to-support-non-DMA-capable-SPI-ctrl.patch
+	if [ -f arch/riscv/configs/mpfs_defconfig ] ; then
+		patch -p1 < ../patches/linux/0001-Add-BeagleV-Fire-device-tree.patch
+		patch -p1 < ../patches/linux/0001-PCIe-Change-controller-and-bridge-base-address.patch
+		patch -p1 < ../patches/linux/0001-GPIO-Add-Microchip-CoreGPIO-driver.patch
+		patch -p1 < ../patches/linux/0001-ADC-Add-Microchip-MCP356X-driver.patch
+		patch -p1 < ../patches/linux/0001-Microchip-QSPI-Add-regular-transfers.patch
+		patch -p1 < ../patches/linux/0001-BeagleV-Fire-Add-printk-to-IM219-driver-for-board-te.patch
+		patch -p1 < ../patches/linux/0001-MMC-SPI-Hack-to-support-non-DMA-capable-SPI-ctrl.patch
+	fi
 	touch .patched
 fi
 
-cp -v ../patches/linux/Makefile arch/riscv/boot/dts/microchip/
+if [ -f arch/riscv/configs/mpfs_defconfig ] ; then
+	cp -v ../patches/linux/Makefile arch/riscv/boot/dts/microchip/Makefile
+	cp -v ../patches/linux/dts/mpfs-beaglev-fire.dts arch/riscv/boot/dts/microchip/
+	cp -v ../patches/linux/dts/mpfs-beaglev-fire-fabric.dtsi arch/riscv/boot/dts/microchip/
+else
+	cp -v ../patches/linux/mainline/Makefile arch/riscv/boot/dts/microchip/Makefile
+	cp -v ../patches/linux/mainline/dts/mpfs-beaglev-fire.dts arch/riscv/boot/dts/microchip/
+	cp -v ../patches/linux/mainline/dts/mpfs-beaglev-fire-fabric.dtsi arch/riscv/boot/dts/microchip/
+fi
 
 make ARCH=riscv CROSS_COMPILE=${CC} clean
-make ARCH=riscv CROSS_COMPILE=${CC} mpfs_defconfig
+
+if [ -f arch/riscv/configs/mpfs_defconfig ] ; then
+	make ARCH=riscv CROSS_COMPILE=${CC} mpfs_defconfig
+else
+	make ARCH=riscv CROSS_COMPILE=${CC} defconfig
+fi
 
 #
 # Scheduler features
@@ -134,7 +149,15 @@ tar --create --gzip --file "../${KERNEL_UTS}-modules.tar.gz" ./*
 cd "${wdir}/linux/" || exit
 rm -rf "${wdir}/deploy/tmp" || true
 
-cp -v ./.config ../patches/linux/mpfs_defconfig
+if [ -f arch/riscv/configs/mpfs_defconfig ] ; then
+	cp -v ./.config ../patches/linux/mpfs_defconfig
+	cp -v ./arch/riscv/boot/dts/microchip/mpfs-beaglev-fire.dts ../patches/linux/dts/mpfs-beaglev-fire.dts
+	cp -v ./arch/riscv/boot/dts/microchip/mpfs-beaglev-fire-fabric.dtsi ../patches/linux/dts/mpfs-beaglev-fire-fabric.dtsi
+else
+	cp -v ./.config ../patches/linux/mainline/defconfig
+	cp -v ./arch/riscv/boot/dts/microchip/mpfs-beaglev-fire.dts ../patches/linux/mainline/dts/mpfs-beaglev-fire.dts
+	cp -v ./arch/riscv/boot/dts/microchip/mpfs-beaglev-fire-fabric.dtsi ../patches/linux/mainline/dts/mpfs-beaglev-fire-fabric.dtsi
+fi
 if [ ! -d ../deploy/input/ ] ; then
 	mkdir -p ../deploy/input/ || true
 fi
